/!\ RUN IN PYTHON ONLY /!\
This is a Python file disguised as a Bitmap file. It might cause severe issues if ran in an image viewer. Figure out how to open it in Python, it was made this way, and serves no purpose anymore. It was a reminder to use a USB drive for the software.